requirements.txt
knack
packaging
azure-cli-core
azure-cli